refactor payload building into separate file, send empty payload for puzzles banner if user is an existing supporter
